+++ sys/cpu/x86_64/include/atomic.h
@@ -158,7 +158,7 @@ atomic_readandclear_long(volatile u_long *addr)
 static __inline u_int
 atomic_readandclear_int(volatile u_int *addr)
 {
-	u_long res;
+	u_int res;
 
 	res = 0;
 	__asm __volatile(
